Update getting-started.rst
Also this paragraph is difficult to follow :-
Note

You may be wondering why there is a Withdraw method at all. Why not just transfer the funds to their owner as soon as they become available? We do this because the owner of these funds could be another contract that we don't control. In sending the funds back, we would potentially be calling unknown computation using another user's funds. As such, this Withdraw pattern is really common in smart contracts. If users call a contract, that contract execution should never delve into an untrusted contract's code.